+++ types/models.d.ts
@@ -71,6 +71,7 @@ declare module 'mongoose' {
     ordered?: boolean;
     lean?: boolean;
     throwOnValidationError?: boolean;
+    timestamps?: boolean | QueryTimestampsConfig;
   }
 
   interface InsertManyResult<T> extends mongodb.InsertManyResult<T> {
@@ -190,19 +191,27 @@ declare module 'mongoose' {
     mongodb.InsertOneModel<TSchema> & MongooseBulkWritePerOperationOptions;
 
   export type ReplaceOneModel<TSchema extends mongodb.Document = mongodb.Document> =
-    mongodb.ReplaceOneModel<TSchema> & MongooseBulkWritePerOperationOptions;
+    Omit<mongodb.ReplaceOneModel<TSchema>, 'filter'> &
+    { filter: QueryFilter<TSchema> } &
+    MongooseBulkUpdatePerOperationOptions;
 
   export type UpdateOneModel<TSchema extends mongodb.Document = mongodb.Document> =
-    mongodb.UpdateOneModel<TSchema> & MongooseBulkUpdatePerOperationOptions;
+    Omit<mongodb.UpdateOneModel<TSchema>, 'filter'> &
+    { filter: QueryFilter<TSchema> } &
+    MongooseBulkUpdatePerOperationOptions;
 
   export type UpdateManyModel<TSchema extends mongodb.Document = mongodb.Document> =
-    mongodb.UpdateManyModel<TSchema> & MongooseBulkUpdatePerOperationOptions;
+    Omit<mongodb.UpdateManyModel<TSchema>, 'filter'> &
+    { filter: QueryFilter<TSchema> } &
+    MongooseBulkUpdatePerOperationOptions;
 
   export type DeleteOneModel<TSchema extends mongodb.Document = mongodb.Document> =
-    mongodb.DeleteOneModel<TSchema>;
+    Omit<mongodb.DeleteOneModel<TSchema>, 'filter'> &
+    { filter: QueryFilter<TSchema> };
 
   export type DeleteManyModel<TSchema extends mongodb.Document = mongodb.Document> =
-    mongodb.DeleteManyModel<TSchema>;
+    Omit<mongodb.DeleteManyModel<TSchema>, 'filter'> &
+    { filter: QueryFilter<TSchema> };
 
   export type AnyBulkWriteOperation<TSchema extends mongodb.Document = mongodb.Document> =
     | { insertOne: InsertOneModel<TSchema> }